KAN-16 refactor: move sidebar to widgets
